stream/requirements.txt
alpaca-py
kafka-python
duckdb
requests
python-dotenv
pandas 
matplotlib 
mplfinance
yfinance 
tqdm 
pandas 
pyarrow 
matplotlib 
seaborn
